Move data source selectable to its own folder and fix test
