+++ src/plugins/data_source_management/public/components/data_source_selectable/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+export { DataSourceSelectable } from './data_source_selectable';
